requirements.txt
openai
pandas
numpy
scipy
sklearn
networkx
langchain
torch
dgl
transformers
python-igraph
matplotlib
mpmath
